requirements.txt
click==6.7
Flask==0.12.2
Flask-Login==0.4.0
Flask-Script==2.0.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
progressbar2==3.34.3
python-utils==2.2.0
six==1.11.0
Werkzeug==0.12.2
